refactor: extract shared TodayCategoryLabel, fix timezone bug, dedupe load-more posts
- Extract duplicated CategoryLabel into shared TodayCategoryLabel component
- Fix Caracas cutoff date using Intl.DateTimeFormat instead of host-offset math
- Rename date helper and add comment explaining 3-day window intent
- Add rounding invariant comment to getSecondaryPosts
- Remove meaningless key prop from single TodayHeroPost article roots
- Deduplicate appended posts in useCategoryPosts mutate to fix duplicate key warning
