+++ src/components/TodayHeroPost/index.tsx
@@ -3,36 +3,14 @@
 import { TodayPost } from '@app/actions/getTodayYesterdayPosts'
 import { DateTime } from '@components/DateTime'
 import { HoverPrefetchLink } from '@components/HoverPrefetchLink'
-import { getPostCategoriesClasses } from '@components/PostCategories/styles'
-import { processCategories } from '@lib/utils/processCategories'
+import { TodayCategoryLabel } from '@components/TodayCategoryLabel'
 import { GAEvent } from '@lib/utils/ga'
-import { GA_EVENTS, CATEGORY_PATH } from '@lib/constants'
+import { GA_EVENTS } from '@lib/constants'
 import { limitStringCharacters } from '@lib/utils/limitStringCharacters'
-import Link from 'next/link'
 
 const WIDE_IMAGE_THRESHOLD = 1200
 
-const CategoryLabel = ({
-  categories,
-  className
-}: {
-  categories: TodayPost['categories']
-  className?: string
-}) => {
-  const processed = processCategories(categories.edges, 1)
-  if (!processed.length) return null
-  const cat = processed[0]?.node
-  const classes = getPostCategoriesClasses(className)
-  if (!cat) return null
-  return (
-    <Link href={`${CATEGORY_PATH}/${cat.slug}`} className={classes}>
-      {cat.name}
-    </Link>
-  )
-}
-
 const TodayHeroPost = ({
-  id,
   title,
   uri,
   date,
@@ -50,10 +28,7 @@ const TodayHeroPost = ({
   // Wide image: full-width overlay, no horizontal padding (bleeds to container edges)
   if (isWide) {
     return (
-      <article
-        key={id}
-        className='relative -mx-6 mb-4 overflow-hidden sm:-mx-7 sm:rounded-sm'
-      >
+      <article className='relative -mx-6 mb-4 overflow-hidden sm:-mx-7 sm:mx-0 sm:rounded-sm'>
         <div
           className='relative w-full overflow-hidden'
           style={{ maxHeight: 500 }}
@@ -92,7 +67,7 @@ const TodayHeroPost = ({
           />
           <div className='absolute right-0 bottom-0 left-0 p-4 md:p-6'>
             {categories && (
-              <CategoryLabel
+              <TodayCategoryLabel
                 categories={categories}
                 className='mb-2 text-white/90'
               />
@@ -127,7 +102,7 @@ const TodayHeroPost = ({
 
   // Smaller image: image left, text right
   return (
-    <article key={id} className='mt-6 mb-4 flex flex-col gap-5 sm:flex-row'>
+    <article className='mt-6 mb-4 flex flex-col gap-5 sm:flex-row'>
       {featuredImage && (
         <div className='relative w-full shrink-0 overflow-hidden sm:w-[55%]'>
           <HoverPrefetchLink
@@ -159,7 +134,7 @@ const TodayHeroPost = ({
       )}
       <div className='flex flex-col justify-center'>
         {categories && (
-          <CategoryLabel
+          <TodayCategoryLabel
             categories={categories}
             className='text-primary mb-1 uppercase'
           />
