+++ src/components/TodayNewsCard/index.tsx
@@ -3,31 +3,10 @@
 import { TodayPost } from '@app/actions/getTodayYesterdayPosts'
 import { DateTime } from '@components/DateTime'
 import { HoverPrefetchLink } from '@components/HoverPrefetchLink'
-import { getPostCategoriesClasses } from '@components/PostCategories/styles'
-import { processCategories } from '@lib/utils/processCategories'
+import { TodayCategoryLabel } from '@components/TodayCategoryLabel'
 import { GAEvent } from '@lib/utils/ga'
-import { GA_EVENTS, CATEGORY_PATH } from '@lib/constants'
+import { GA_EVENTS } from '@lib/constants'
 import { limitStringCharacters } from '@lib/utils/limitStringCharacters'
-import Link from 'next/link'
-
-const CategoryLabel = ({
-  categories,
-  className
-}: {
-  categories: TodayPost['categories']
-  className?: string
-}) => {
-  const processed = processCategories(categories.edges, 1)
-  if (!processed.length) return null
-  const cat = processed[0]?.node
-  const classes = getPostCategoriesClasses(className)
-  if (!cat) return null
-  return (
-    <Link href={`${CATEGORY_PATH}/${cat.slug}`} className={classes}>
-      {cat.name}
-    </Link>
-  )
-}
 
 const TodayNewsCard = ({
   id,
@@ -74,7 +53,7 @@ const TodayNewsCard = ({
         )}
       </HoverPrefetchLink>
       {categories && (
-        <CategoryLabel
+        <TodayCategoryLabel
           categories={categories}
           className='text-primary pt-1 pb-2 uppercase'
         />
